Create MIMEType file and move logic out of DownloadHelper
